Moved menu-building code to a menu.mjs file.
Renamed dialogHandler to onMenuItemClick. Seems more honest now that I've simplified the menu building code.

Moved type definitions to a _types file.

Removed dnd-menu-config CSS.
