+++ styles/menu-config.css
@@ -1,152 +1,3 @@
-#dnd-menu-config .menu-config {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  padding: 10px;
-}
-
-#dnd-menu-config .menu-item,
-#dnd-menu-config .menu-label {
-  font-weight: bold;
-}
-
-#dnd-menu-config .menu-item {
-  padding: 6px 12px;
-  font-size: 0.9em;
-}
-
-#dnd-menu-config .menu-item:hover {
-  transform: scale(1.02);
-}
-
-#dnd-menu-config .menu-label {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  gap: 8px;
-  font-size: 14px;
-  padding-top: 5px;
-}
-
-#dnd-menu-config .menu-name {
-  flex-grow: 1;
-}
-
-#dnd-menu-config input[type="checkbox"] {
-  display: none;
-}
-
-#dnd-menu-config .toggle-indicator {
-  position: relative;
-  width: 40px;
-  height: 20px;
-  background: transparent;
-  border: 1px dotted var(--dnd5e-color-gold);
-  border-radius: 20px;
-  transition: background 0.3s ease;
-}
-
-#dnd-menu-config .toggle-indicator::before {
-  content: "";
-  position: absolute;
-  width: 16px;
-  height: 16px;
-  background: var(--dnd5e-color-gold);
-  border-radius: 50%;
-  top: 1px;
-  left: 2px;
-  transition: transform 0.3s ease, background 0.3s ease;
-}
-
-#dnd-menu-config input[type="checkbox"]:checked+.toggle-indicator {
-  background: var(--dnd5e-color-gold);
-}
-
-#dnd-menu-config input[type="checkbox"]:checked+.toggle-indicator::before {
-  transform: translateX(20px);
-  background: white;
-}
-
-#dnd-menu-config hr {
-  margin-block: 5px;
-  border: none;
-  height: 1px;
-  background: linear-gradient(to right,
-      transparent,
-      var(--dnd5e-color-gold),
-      transparent);
-}
-
-#dnd-menu-config hr.ampersand {
-  position: relative;
-  height: 20px;
-  background: url("./assets/ampersand-gold.svg") no-repeat center / contain;
-  margin: 0;
-  margin-bottom: -4px;
-}
-
-#dnd-menu-config hr.ampersand::before,
-#dnd-menu-config hr.ampersand::after {
-  content: "";
-  position: absolute;
-  inset-block: 9.5px;
-  inset-inline: 0;
-}
-
-#dnd-menu-config hr.ampersand::before {
-  background: linear-gradient(to left, var(--dnd5e-color-gold), transparent);
-  inset-inline-end: calc(50% + 14px);
-}
-
-#dnd-menu-config hr.ampersand::after {
-  background: linear-gradient(to right, var(--dnd5e-color-gold), transparent);
-  inset-inline-start: calc(50% + 14px);
-}
-
-#dnd-menu-config .menu-header {
-  display: flex;
-  align-items: flex-start;
-  gap: 10px;
-  font-family: "Modesto Condensed", "Palatino Linotype", serif;
-  font-size: 2rem;
-  font-weight: bold;
-}
-
-#dnd-menu-config .menu-header img {
-  align-self: flex-start;
-  border: none;
-}
-
-#dnd-menu-config .title-description {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-#dnd-menu-config .title-description h2 {
-  margin: 0;
-  font-size: 2.25rem;
-  font-family: "Modesto Condensed", "Palatino Linotype", serif;
-  font-weight: bold;
-  line-height: 1;
-  border-bottom: none;
-}
-
-#dnd-menu-config .title-description p {
-  margin: 0;
-  font-size: 1rem;
-  font-family: "Signika", sans-serif;
-}
-
-.theme-light #dnd-menu-config .title-description p {
-  color: #4b4a44;
-}
-
-#dnd-menu-config .title-description p {
-  color: #999;
-}
-
-#dnd-menu-config button[type="submit"],
 #reference-dialog-v2 button[type="submit"] {
   width: 100%;
   margin-top: 10px;
